requirements.txt
absl-py #2.3.1
aiohappyeyeballs #2.6.1
aiohttp #3.12.15
aiosignal #1.4.0
annotated-types #0.7.0
anyascii #0.3.3
anyio #4.11.0
attrs #25.3.0
audioread #3.0.1
av #16.0.1
babel #2.17.0
backoff #2.2.1
bangla #0.0.5
bcrypt #5.0.0
beautifulsoup4 #4.13.5
blinker #1.9.0
blis #1.2.1
build #1.3.0
cachetools #5.5.2
catalogue #2.0.10
certifi #2025.8.3
cffi #2.0.0
charset-normalizer #3.4.3
chromadb #1.1.0
click #8.3.0
cloudpathlib #0.23.0
coloredlogs #15.0.1
confection #0.1.5
contourpy #1.3.3
coqpit #0.0.17
ctranslate2 #4.6.0
cycler #0.12.1
cymem #2.0.11
Cython #3.1.4
dataclasses-json #0.6.7
dateparser #1.1.8
decorator #5.2.1
distro #1.9.0
docopt #0.6.2
durationpy #0.10
einops #0.8.1
elevenlabs #2.18.0
encodec #0.1.1
faiss-cpu #1.12.0
fastapi #0.117.1
faster-whisper #1.2.0
filelock #3.19.1
Flask #3.1.2
flatbuffers #25.9.23
fonttools #4.60.1
frozenlist #1.7.0
fsspec #2025.9.0
g2pkk #0.1.2
google-auth #2.40.3
googleapis-common-protos #1.70.0
grpcio #1.75.1
gruut #2.2.3
gruut-ipa #0.13.0
gruut_lang_de #2.0.1
gruut_lang_en #2.0.1
gruut_lang_es #2.0.1
gruut_lang_fr #2.0.2
h11 #0.16.0
hangul-romanize #0.1.0
hf-xet #1.1.10
httpcore #1.0.9
httptools #0.6.4
httpx #0.28.1
httpx-sse #0.4.1
huggingface-hub #0.35.1
humanfriendly #10.0
idna #3.10
importlib_metadata #8.7.0
importlib_resources #6.5.2
inflect #7.5.0
itsdangerous #2.2.0
jamo #0.4.1
jieba #0.42.1
Jinja2 #3.1.6
jiter #0.11.0
joblib #1.5.2
jsonlines #1.2.0
jsonpatch #1.33
jsonpointer #3.0.0
jsonschema #4.25.1
jsonschema-specifications #2025.9.1
kiwisolver #1.4.9
kubernetes #33.1.0
langchain #0.3.27
langchain-community #0.3.29
langchain-core #0.3.79
langchain-openai #0.3.35
langchain-text-splitters #0.3.11
langcodes #3.5.0
langdetect #1.0.9
langsmith #0.4.31
language_data #1.3.0
lazy_loader #0.4
librosa #0.11.0
llvmlite #0.45.1
marisa-trie #1.3.1
Markdown #3.9
markdown-it-py #4.0.0
MarkupSafe #3.0.2
marshmallow #3.26.1
matplotlib #3.10.7
mdurl #0.1.2
mmh3 #5.2.0
more-itertools #10.8.0
mpmath #1.3.0
msgpack #1.1.2
multidict #6.6.4
murmurhash #1.0.13
mypy_extensions #1.1.0
networkx #2.8.8
nltk #3.9.2
num2words #0.5.14
numba #0.62.1
numpy #1.26.4
oauthlib #3.3.1
onnxruntime #1.23.0
openai #1.109.1
packaging #25.0
pandas #1.5.3
pillow #11.3.0
platformdirs #4.5.0
pooch #1.8.2
posthog #5.4.0
preshed #3.0.10
propcache #0.3.2
protobuf #6.32.1
psutil #7.1.0
pyasn1 #0.6.1
pyasn1_modules #0.4.2
pybase64 #1.4.2
pycparser #2.23
pydantic #2.11.9
pydantic-settings #2.11.0
pydantic_core #2.33.2
Pygments #2.19.2
pynndescent #0.5.13
pyparsing #3.2.5
PyPika #0.48.9
pypinyin #0.55.0
pyproject_hooks #1.2.0
pysbd #0.3.4
python-crfsuite #0.9.11
python-dateutil #2.9.0.post0
python-dotenv #1.1.1
pytz #2025.2
PyYAML #6.0.2
referencing #0.36.2
regex #2025.9.18
requests #2.32.5
requests-oauthlib #2.0.0
requests-toolbelt #1.0.0
rich #14.1.0
rpds-py #0.27.1
rsa #4.9.1
safetensors #0.6.2
scikit-learn #1.7.2
scipy #1.16.2
sentence-transformers #5.1.1
shellingham #1.5.4
six #1.17.0
smart_open #7.3.1
sniffio #1.3.1
soundfile #0.13.1
soupsieve #2.8
soxr #1.0.0
spacy #3.8.7
spacy-legacy #3.0.12
spacy-loggers #1.0.5
SQLAlchemy #2.0.43
srsly #2.5.1
starlette #0.48.0
tensorboard #2.20.0
tensorboard-data-server #0.7.2
thinc #8.3.4
threadpoolctl #3.6.0
tiktoken #0.12.0
tokenizers #0.22.1
transformers #4.56.2
# TTS #0.22.0
# urllib3 #2.5.0
# uvicorn #0.37.0
# uvloop #0.21.0
# wasabi #1.1.3
# watchfiles #1.1.0
# weasel #0.4.1
# websocket-client #1.8.0
# websockets #15.0.1
# Werkzeug #3.1.3
# wrapt #1.17.3
# yarl #1.20.1
# zipp #3.23.0
# zstandard #0.25.0
